Added team The Triple Entente.md
The Triple Entente (from French entente [ɑ̃tɑ̃t] meaning "friendship, understanding, agreement") describes the informal understanding between the Russian Empire, the French Third Republic, and the United Kingdom of Great Britain and Ireland. It was built upon the Franco-Russian Alliance of 1894, the Entente Cordiale of 1904 between France and Britain, and the Anglo-Russian Entente of 1907. The Triple Entente, unlike the Triple Alliance or the Franco-Russian Alliance itself, was not an alliance of mutual defense.